Refactor to V1
