+++ safe_transaction_service/history/urls_v2.py
@@ -36,9 +36,4 @@ urlpatterns = [
         views_v2.AllTransactionsListView.as_view(),
         name="all-transactions",
     ),
-    path(
-        "safes/<str:address>/export/",
-        views_v2.SafeExportView.as_view(),
-        name="safe-export",
-    ),
 ]
