+++ safe_transaction_service/history/views_v2.py
@@ -26,7 +26,6 @@ from . import filters, pagination, serializers
 from .cache import CacheSafeTxsView, cache_txs_view_for_address
 from .models import MultisigTransaction, SafeContract, SafeContractDelegate
 from .pagination import DummyPagination
-from .serializers import SafeExportTransactionRequestParams
 from .services import BalanceServiceProvider, TransactionServiceProvider
 from .services.balance_service import Balance
 from .services.collectibles_service import CollectiblesServiceProvider
@@ -633,105 +632,3 @@ class AllTransactionsListView(ListAPIView):
             "W/" + hashlib.md5(str(response.data["results"]).encode()).hexdigest(),
         )
         return response
-
-
-@extend_schema(
-    parameters=[
-        OpenApiParameter(
-            "execution_date__gte",
-            location="query",
-            type=OpenApiTypes.DATETIME,
-            description="Filter transactions executed after this date (ISO format)",
-        ),
-        OpenApiParameter(
-            "execution_date__lte",
-            location="query",
-            type=OpenApiTypes.DATETIME,
-            description="Filter transactions executed before this date (ISO format)",
-        ),
-        OpenApiParameter(
-            "limit",
-            location="query",
-            type=OpenApiTypes.INT,
-            description="Maximum number of transactions to return (max 1000)",
-        ),
-        OpenApiParameter(
-            "offset",
-            location="query",
-            type=OpenApiTypes.INT,
-            description="Number of transactions to skip",
-        ),
-    ],
-    responses={
-        200: OpenApiResponse(
-            response=serializers.SafeExportTransactionSerializer(many=True)
-        ),
-        404: OpenApiResponse(description="Safe not found"),
-        422: OpenApiResponse(
-            description="Safe address checksum not valid",
-            response=serializers.CodeErrorResponse,
-        ),
-    },
-)
-class SafeExportView(GenericAPIView):
-    """
-    Export endpoint for CSV export feature - optimized for large datasets
-    """
-
-    serializer_class = serializers.SafeExportTransactionSerializer
-
-    def get_query_params(self):
-        serializer = SafeExportTransactionRequestParams(
-            data=self.request.query_params.dict()
-        )
-        serializer.is_valid(raise_exception=True)
-
-        validated_data = serializer.validated_data
-        parsed_execution_date_gte = validated_data.get("execution_date__gte")
-        parsed_execution_date_lte = validated_data.get("execution_date__lte")
-        return parsed_execution_date_gte, parsed_execution_date_lte
-
-    def get(self, request, address):
-        """
-        Get transactions optimized for CSV export with transfer information.
-        The maximum limit allowed is 1000.
-        """
-
-        if not fast_is_checksum_address(address):
-            return Response(
-                status=status.HTTP_422_UNPROCESSABLE_ENTITY,
-                data={
-                    "code": 1,
-                    "message": "Checksum address validation failed",
-                    "arguments": [address],
-                },
-            )
-
-        try:
-            SafeContract.objects.get(address=address)
-        except SafeContract.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-
-        # Get query params
-        parsed_execution_date_gte, parsed_execution_date_lte = self.get_query_params()
-
-        # Get pagination
-        paginator = pagination.ListPagination(self.request, max_limit=1000)
-        limit = paginator.get_limit(request)
-        offset = paginator.get_offset(request)
-
-        # Get transactions from service
-        transaction_service = TransactionServiceProvider()
-        transactions, total_count = transaction_service.get_export_transactions(
-            address,
-            execution_date_gte=parsed_execution_date_gte,
-            execution_date_lte=parsed_execution_date_lte,
-            limit=limit,
-            offset=offset,
-        )
-        paginator.set_count(total_count)
-
-        # Serialize the data
-        serializer = self.get_serializer(transactions, many=True)
-
-        return paginator.get_paginated_response(serializer.data)
